refact: create a util class with all methods to select colors and dimens
