+++ app/src/main/java/com/lucasmarciano/composeproject/ui/InterfaceFactory.kt
@@ -3,11 +3,8 @@ package com.lucasmarciano.composeproject.ui
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.width
-import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalDensity
 import androidx.navigation.NavController
 import com.lucasmarciano.composeproject.features.home.components.BlueCardsList
 import com.lucasmarciano.composeproject.features.home.components.CardsList
@@ -17,12 +14,11 @@ import com.lucasmarciano.composeproject.ui.components.Banner
 import com.lucasmarciano.composeproject.ui.components.Profile
 import com.lucasmarciano.composeproject.ui.components.SecondTitle
 import com.lucasmarciano.composeproject.ui.components.Title
-import com.lucasmarciano.composeproject.ui.theme.ColorBlueCard
-import com.lucasmarciano.composeproject.ui.theme.ColorPinkCalifornia
+import com.lucasmarciano.composeproject.ui.utils.getColorValue
+import com.lucasmarciano.composeproject.ui.utils.getDimensionValue
 import com.lucasmarciano.composeproject.ui.values.BannerComponent
 import com.lucasmarciano.composeproject.ui.values.CardBlueListComponent
 import com.lucasmarciano.composeproject.ui.values.CardIconListComponent
-import com.lucasmarciano.composeproject.ui.values.ColorComponent
 import com.lucasmarciano.composeproject.ui.values.HomeTitleComponent
 import com.lucasmarciano.composeproject.ui.values.InterfaceItemComponent
 import com.lucasmarciano.composeproject.ui.values.MenuItemComponent
@@ -40,11 +36,11 @@ fun InterfaceFactory(item: InterfaceItemComponent, navController: NavController)
         }
         Components.MAIN_TITLE -> {
             val homeValue = (item as TitleComponent)
-            Title(homeValue.name, composeColorValue(homeValue.color))
+            Title(homeValue.name, getColorValue(homeValue.color))
         }
         Components.SECOND_TITLE -> {
             val secondTitle = (item as SecondTitleComponent)
-            SecondTitle(secondTitle.name, composeColorValue(secondTitle.color))
+            SecondTitle(secondTitle.name, getColorValue(secondTitle.color))
         }
         Components.CARD_ICON_LIST -> {
             val cardIconList = (item as CardIconListComponent).items
@@ -58,8 +54,8 @@ fun InterfaceFactory(item: InterfaceItemComponent, navController: NavController)
             val spacerComponent = (item as SpacerComponent)
             Spacer(
                 modifier = Modifier
-                    .height(composeDimensionValue(spacerComponent.verticalSize))
-                    .width(composeDimensionValue(spacerComponent.horizontalSize))
+                    .height(getDimensionValue(spacerComponent.verticalSize))
+                    .width(getDimensionValue(spacerComponent.horizontalSize))
             )
         }
         Components.BANNER -> {
@@ -78,22 +74,3 @@ fun InterfaceFactory(item: InterfaceItemComponent, navController: NavController)
         else -> {}
     }
 }
-
-@Composable
-private fun composeColorValue(colorComponent: ColorComponent): Color {
-    return when (colorComponent) {
-        ColorComponent.PINK -> ColorPinkCalifornia
-        ColorComponent.BLUE_CARD_BG -> ColorBlueCard
-        ColorComponent.BACKGROUND -> MaterialTheme.colors.background
-        ColorComponent.SECONDARY -> MaterialTheme.colors.secondary
-        ColorComponent.SURFACE -> MaterialTheme.colors.surface
-        ColorComponent.ON_PRIMARY -> MaterialTheme.colors.onPrimary
-        ColorComponent.ON_SECONDARY -> MaterialTheme.colors.onSecondary
-        ColorComponent.ON_BACKGROUND -> MaterialTheme.colors.onBackground
-        ColorComponent.ON_SURFACE -> MaterialTheme.colors.onSurface
-        else -> MaterialTheme.colors.primary
-    }
-}
-
-@Composable
-private fun composeDimensionValue(dimen: Int) = LocalDensity.current.run { dimen.toDp() }
