+++ app/src/main/java/com/lucasmarciano/composeproject/ui/components/toolbar/CollapseToolbar.kt
@@ -1,0 +1,104 @@
+package com.lucasmarciano.composeproject.ui.components.toolbar
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.BackdropScaffold
+import androidx.compose.material.BackdropValue
+import androidx.compose.material.ExperimentalMaterialApi
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.material.rememberBackdropScaffoldState
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.google.accompanist.insets.statusBarsPadding
+import com.lucasmarciano.composeproject.ui.theme.ComposeProjectTheme
+import com.lucasmarciano.composeproject.ui.utils.Const.APP_BAR_COLLAPSED_HEIGHT
+import com.lucasmarciano.composeproject.ui.utils.Const.BottomSheetShape
+import com.lucasmarciano.composeproject.ui.utils.getBgToolbarColor
+import com.lucasmarciano.composeproject.ui.values.ToolbarComponent
+import com.lucasmarciano.composeproject.ui.values.TypeToolbar
+
+@OptIn(ExperimentalMaterialApi::class)
+@Composable
+internal fun CollapseToolbar(
+    toolbar: ToolbarComponent,
+    onClickBack: (String) -> Unit = { },
+    onClick: () -> Unit = { },
+    header: @Composable () -> Unit,
+    content: @Composable () -> Unit,
+) {
+    val backDropState = rememberBackdropScaffoldState(BackdropValue.Revealed)
+
+    BackdropScaffold(
+        modifier = Modifier.background(getBgToolbarColor(typeToolbar = toolbar.type)),
+        frontLayerShape = BottomSheetShape,
+        frontLayerScrimColor = Color.Unspecified,
+        scaffoldState = backDropState,
+        appBar = {
+            Row(
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.SpaceBetween,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .background(getBgToolbarColor(typeToolbar = toolbar.type))
+                    .statusBarsPadding()
+                    .height(APP_BAR_COLLAPSED_HEIGHT)
+                    .padding(horizontal = 8.dp)
+            ) {
+                BuildBackAction(toolbar, onClickBack)
+                Text(
+                    text = toolbar.title,
+                    textAlign = TextAlign.Center,
+                    style = MaterialTheme.typography.subtitle1,
+                    color = Color.White
+                )
+                BuildAction(toolbar, onClick)
+            }
+        },
+        backLayerContent = header,
+        frontLayerContent = content
+    )
+}
+
+@Preview(showBackground = true, backgroundColor = 0xFFFFFFFF, showSystemUi = true)
+@Composable
+fun CollapseToolBarPreview() {
+    ComposeProjectTheme {
+        CollapseToolbar(
+            toolbar = ToolbarComponent("seu négocio", type = TypeToolbar.BLUE),
+            header = {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(300.dp),
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.Center
+                ) {
+                    Text(
+                        modifier = Modifier.fillMaxWidth(),
+                        text = "header",
+                        textAlign = TextAlign.Center,
+                        color = Color.White
+                    )
+                }
+            }
+        ) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.Center
+            ) {
+                Text(text = "content")
+            }
+        }
+    }
+}
